Refactor styles: rename utils to util, comments to comment, and cards to card; migrate styles to new files
