+++ src/pages/Cards.tsx
@@ -1,8 +1,8 @@
 import { footballTeam } from "../data/footballTeam";
 import { FootballPlayer } from "../types/FootballPlayer";
 // import { TableHeader } from "../types/TableHeader";
-import utilStyles from "../styles/utils.module.css";
-import cardStyles from "../styles/cards.module.css";
+import utilStyles from "../styles/util.module.css";
+import cardStyles from "../styles/card.module.css";
 import { useNavigate } from "react-router-dom";
 
 const playersList: FootballPlayer[] = footballTeam;
